Fix related link desc and move topic titles to correct stylesheet
